Refactored cli api build, so that it can be re-used without local
