+++ version/version.go
@@ -2,7 +2,7 @@ package version
 
 var (
 	// VERSION should be updated by hand at each release
-	VERSION = "0.2.0-dev"
+	VERSION = "0.3.0-dev"
 
 	// GITCOMMIT will be overwritten automatically by the build system
 	GITCOMMIT = "HEAD"
